made window integer

+++ lgndbdt/extraction_utils/Extraction.py
@@ -87,7 +87,7 @@ def extraction(paramArr, paramKeys):
         windowTail       = tailWindow(pa["tp_0"][i], pa["dt"][i])
         if windowTail[0] + 250 >= len(wfCorr[i])-100:
             print(windowTail)
-            windowTail[0]    = windowTail[0] + np.floor((len(wfCorr[i])-100-windowTail[0])/5)
+            windowTail[0]    = windowTail[0] + int(np.floor((len(wfCorr[i])-100-windowTail[0])/5))
             print(windowTail)
         else:
             windowTail[0]    = windowTail[0] + 250
